services/ocr/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
numpy>=1.24
opencv-python>=4.9.0
pillow>=10.0.0
ultralytics>=8.2.0
transformers>=4.41.0
torch>=2.2.0
torchvision>=0.17.0
python-dotenv>=1.0.1
pdf2image>=1.17.0
mysql-connector-python>=8.0.0
python-multipart>=0.0.6
llmwhisperer-client>=2.0.0
